Add Camel Cards Game logic and unit tests
This commit includes the creation of new classes necessary for a game of Camel cards, including different types of hands (normal and wild card), game logic to calculate winnings, and unit tests. It also modifies the DayRunner project to include Day07 execution.

Why is this change necessary?
This game and its logic is being added to the suite of games available in DayRunner, presenting another option for users to enjoy.