Add ControlPanel to find max energized tiles and update tests
Extended the functionality of the application by adding a ControlPanel class to search and count the maximum number of energized tiles. Altered Contraption constructor to accept starting Beam as a parameter. Beam and Direction classes have been modified to be publicly accessible. Updated existing tests and added new tests for the ControlPanel class.