Add PathFinder functionality and tests
This commit introduces a new 'PathFinder' class that allows calculating the shortest path based on a set of input criteria. Using methods that calculate heuristic, retrieve neighbours of a node, and find a path, the class provides advanced path-finding capabilities. Furthermore, the commit also includes new tests for the 'PathFinder' class and introduces an update to the project's solution file, including additions of new 'Day17' project files and 'Day17.Tests' project files.